requirements.txt
# Core RAG Framework
sentence-transformers==2.2.2
chromadb==0.4.15
langchain==0.1.0
langchain-community==0.0.13

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.9.0
pymupdf==1.23.8
pytesseract==0.3.10

# Machine Learning & NLP
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
accelerate==0.24.1
bitsandbytes==0.41.3
optimum==1.16.0
numpy==1.24.3
scikit-learn==1.3.0
spacy==3.7.2

# LLM Integration - Cloud Providers
openai==1.3.0
anthropic==0.7.7

# LLM Integration - Local Models
ollama==0.1.7
llama-cpp-python==0.2.11
tokenizers==0.15.0
safetensors==0.4.1

# Data Processing
pandas==2.0.3
regex==2023.8.8
python-dotenv==1.0.0

# Web Interface (Optional)
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
plotly==5.17.0

# Utilities
tqdm==4.66.1
pydantic==2.5.0
typing-extensions==4.8.0
psutil==5.9.6
GPUtil==1.4.0 